@@ -6,7 +6,9 @@ import {
   SiSlack,
   SiVercel,
   SiNetlify,
-  SiMacos
+  SiMacos,
+  SiXcode,
+  SiGithub
 } from "react-icons/si";
 
 function Toolstack() {
@@ -19,13 +21,13 @@ function Toolstack() {
         <SiVisualstudiocode />
       </Col>
       <Col xs={4} md={2} className="tech-icons">
+        <SiXcode />
+      </Col>
+      <Col xs={4} md={2} className="tech-icons">
         <SiPostman />
       </Col>
       <Col xs={4} md={2} className="tech-icons">
-        <SiNetlify />
-      </Col>
-      <Col xs={4} md={2} className="tech-icons">
-        <SiVercel />
+        <SiGithub />
       </Col>
     </Row>
   );
