@@ -1,5 +1,6 @@
-import React from "react";
-import { Container, Row, Col } from "react-bootstrap";
+import React, { useState } from "react";
+import { Container, Row, Col, ButtonGroup, Button } from "react-bootstrap";
+import { motion } from "framer-motion";
 import ProjectCard from "./ProjectCards";
 import Particle from "../Particle";
 import hitech from "../../Assets/Projects/hitech.jpeg";
@@ -12,94 +13,157 @@ import sirekam from "../../Assets/Projects/sirekam.png";
 import arch from "../../Assets/Projects/arch.png";
 
 function Projects() {
+  const [filter, setFilter] = useState("All");
+
+  const projects = [
+    {
+      imgPath: hitech,
+      title: "Online Presence With Face Recognition and Geolocation",
+      description: "I’m building a website for the Hi-Technology 2023 Software Fair. The tech stack I used includes SvelteKit, Supabase, and PostgreSQL. For face recognition, I utilized the face-api.js API, and for geolocation, I implemented Mozilla Web API.",
+      category: "Web",
+      ghLink: "https://github.com/RickyPrima30/presensi-facerecognition-geolocation",
+      demoLink: "https://presensi-mhs.netlify.app/",
+    },
+    {
+      imgPath: lms,
+      title: "Learning Management System",
+      description: "Developed a learning management system website for a mini project organized by an association. Integrated OAuth, email verification link, and email verification code. The platform allows users to engage with structured learning materials in video format. The tech stack includes Next.js, PostgreSQL, Supabase, Prisma, and Shadcnui.",
+      category: "Web",
+      ghLink: "https://github.com/rickyprimay/lms",
+      demoLink: "https://rickieslearn.vercel.app/",
+    },
+    {
+      imgPath: codejam,
+      title: "Website Code Jam Academy",
+      description: "I’m building a website for Code Jam Academy 2023. The project is an internal training website by Himpunan Mahasiswa Teknik Informatika. The theme of the project is Laravel 10 CRUD with Authentication. Tech stack: HTML5 and TailwindCSS.",
+      category: "Web",
+      ghLink: "https://github.com/codejamacademy/codejamacademy.github.io",
+      demoLink: "https://codejamacademy.github.io",
+    },
+    {
+      imgPath: ppko,
+      title: "jARictku",
+      description: "I’m building this website for the Program Penguatan Kapasitas Ormawa (PPK-O). I served as a developer on the team. This project features augmented reality technology using Laravel, TailwindCSS, MySQL, and AR.js.",
+      category: "Web",
+      demoLink: "https://jaricktu.com",
+    },
+    {
+      imgPath: ticketify,
+      title: "Ticketify",
+      description: "Working on the development of Ticketify, an event ticketing website similar to Tiket.com and YesPlis. The website is still in the development phase. Tech stack includes Laravel for backend, React for frontend, and Midtrans as the payment gateway service.",
+      category: "Web",
+      demoLink: "https://ticketify.id",
+    },
+    {
+      imgPath: sirekam,
+      title: "Sirekampolkesyogya",
+      description: "Developed a website, sirekampolkesyogya.com, as a freelance project for students at Poltekkes Yogyakarta to participate in PKM competitions. The website helps record and manage student activity scores.",
+      category: "Web",
+      demoLink: "https://sirekampolkesyogya.com",
+    },
+    {
+      imgPath: arch,
+      title: "Sumberdaya Fakultas Rekaya Industri",
+      description: "Developed a website, sumberdayafri.my.id, as a freelance project for students at Telkom University to complete their Thesis. The website serves as an information and resource platform for faculty students.",
+      category: "Web",
+      demoLink: "https://sumberdayafri.my.id",
+    },
+    {
+      imgPath: tf,
+      title: "Traffic Sense",
+      description: "Participated in the DINACOM software competition by creating Traffic Sense, an app designed to find routes with minimal pollution. It uses Graphhopper API for maps and IoT with ESP-32 and MQ-135 sensors to monitor air quality. I developed the API for the mobile application.",
+      category: "Mobile",
+      demoLink: "https://smart-traffic.my.id/",
+    },
+  ];
+
+  const filteredProjects = projects.filter(
+    (project) => filter === "All" || project.category === filter
+  );
+
   return (
-    <Container fluid className="project-section">
+    <Container
+      fluid
+      className="project-section"
+      style={{
+        background: "#1a1a2e",
+        color: "#fff",
+        padding: "40px 20px",
+      }}
+    >
       <Particle />
       <Container>
-        <h1 className="project-heading">
-          My Recent <strong className="purple">Project </strong>
-        </h1>
-        <p style={{ color: "white" }}>
-          Here are a few projects I've worked on recently and Production.
-        </p>
-        <Row style={{ justifyContent: "center", paddingBottom: "10px" }}>
-          
+        <motion.h1
+          className="project-heading"
+          style={{
+            textAlign: "center",
+            fontWeight: "700",
+            color: "#ffc107",
+            marginBottom: "20px",
+          }}
+          initial={{ opacity: 0, y: -20 }}
+          animate={{ opacity: 1, y: 0 }}
+          transition={{ duration: 0.6 }}
+        >
+          My Recent <strong style={{ color: "#cd5ff8" }}>Projects</strong>
+        </motion.h1>
+        <motion.p
+          style={{ textAlign: "center", marginBottom: "30px" }}
+          initial={{ opacity: 0 }}
+          animate={{ opacity: 1 }}
+          transition={{ delay: 0.3, duration: 0.6 }}
+        >
+          Here are a few projects I've worked on recently and in production.
+        </motion.p>
 
-          <Col md={4} className="project-card">
-            <ProjectCard
-              imgPath={hitech}
-              isBlog={false}
-              title="Online Presence With Face Recognition and Geolocation"
-              description="Im Building Website to joining Hi-Technology 2023 Software Fair, TechStack i use is Svelte Kit, Supabase, PostgreSQL. For Face Recognition i use API from face-api.js and geolocation i use mozila WEB API"
-              ghLink = "https://github.com/RickyPrima30/presensi-facerecognition-geolocation"
-              demoLink = "https://presensi-mhs.netlify.app/"
-            />
-          </Col>
-          <Col md={4} className="project-card">
-            <ProjectCard
-              imgPath={lms}
-              isBlog={false}
-              title="Learning Management System"
-              description="Developed a learning management system website for a mini project organized by the organization. Integrated OAuth, email verification link, and email verification code. The website serves as a learning platform, enabling users to engage with structured learning videos segmented per material. Utilized the following TechStack: Next.js, PostgreSQL, Supabase, Prisma, and Shadcnui."
-              ghLink = "https://github.com/rickyprimay/lms"
-              demoLink = "https://rickieslearn.vercel.app/"
-            />
-          </Col>
-          <Col md={4} className="project-card">
-            <ProjectCard
-              imgPath={codejam}
-              isBlog={false}
-              title="Website Code Jam Academy"
-              description = "Im building this website to prepare Code Jam Academy 2023, Code Jam Academy is internal training website from Himpunan Mahasiswa Teknik Informatika The theme I took was Laravel 10 CRUD and Authentication. I use TechStack HTML5 and TailwindCSS"
-              ghLink = "https://github.com/codejamacademy/codejamacademy.github.io"
-              demoLink = "https://codejamacademy.github.io"
-            />
-          </Col>
-          <Col md={4} className="project-card">
-            <ProjectCard
-              imgPath={ppko}
-              isBlog={false}
-              title="jARictku"
-              description = "Im building this Website to joining Program Penguatan Kapasitas Ormawa(PPK-O), where I served as a developer on the team. This website is a website that has augmented reality technology where I use TechStack Laravel, Tailwindcss, Mysql, and Ar.js as the augmented reality"
-              demoLink = "https://jaricktu.com"
-            />
-          </Col>
-          <Col md={4} className="project-card">
-            <ProjectCard
-              imgPath={ticketify}
-              isBlog={false}
-              title="Ticketify"
-              description = "Working on the development of Ticketify, an event ticketing website similar to Tiket.com and YesPlis, designed for booking event tickets. The website is still in the development/local phase. For this project, I am using Laravel as the backend framework, React for the frontend development, and integrating Midtrans as the payment gateway service."
-              demoLink = "https://ticketify.id"
-            />
-          </Col>
-          <Col md={4} className="project-card">
-            <ProjectCard
-              imgPath={sirekam}
-              isBlog={false}
-              title="Sirekampolkesyogya"
-              description = "Developing a website called sirekampolkesyogya.com as a freelance project for students at Poltekkes Yogyakarta to participate in PKM competitions. This website functions to record and manage the activity scores of the students."
-              demoLink = "https://sirekampolkesyogya.com"
-            />
-          </Col>
-          <Col md={4} className="project-card">
-            <ProjectCard
-              imgPath={arch}
-              isBlog={false}
-              title="Sumberdaya Fakultas Rekaya Industri"
-              description = "Developing a website called sumberdayafri.my.id as a freelance project for students at Telkom University to completed Thesis."
-              demoLink = "https://sumberdayafri.my.id"
-            />
-          </Col>
-          <Col md={4} className="project-card">
-            <ProjectCard
-              imgPath={tf}
-              isBlog={false}
-              title="Traffic Sense"
-              description = "I took part in a software competition called DINACOM and I made an application called Traffic Sense where the purpose of this application was to find routes with little pollution where for the map I used the API from Graphooper and to find out the air quality I made IoT with the ESP-32 module and MQ-135 Water Quality Sensor. where in this project I was tasked with creating an API for Mobile"
-              demoLink = "https://smart-traffic.my.id/"
-            />
-          </Col>
+        {/* Filter Buttons */}
+        <motion.div
+          style={{ display: "flex", justifyContent: "center", marginBottom: "30px" }}
+          initial={{ opacity: 0 }}
+          animate={{ opacity: 1 }}
+          transition={{ delay: 0.5 }}
+        >
+          <ButtonGroup>
+            {["All", "Web", "Mobile"].map((category) => (
+              <Button
+                key={category}
+                variant={filter === category ? "dark" : "light"}
+                onClick={() => setFilter(category)}
+                style={{
+                  border: "2px solid #a24dd386",
+                  color: filter === category ? "#fff" : "#a24dd386",
+                  background: filter === category ? "#a24dd386" : "transparent",
+                  fontWeight: "600",
+                  margin: "0 5px",
+                }}
+              >
+                {category}
+              </Button>
+            ))}
+          </ButtonGroup>
+        </motion.div>
+
+        {/* Project Cards */}
+        <Row style={{ justifyContent: "center" }}>
+          {filteredProjects.map((project, index) => (
+            <Col md={4} key={index} className="mb-4">
+              <motion.div
+                initial={{ opacity: 0, scale: 0.9 }}
+                whileInView={{ opacity: 1, scale: 1 }}
+                viewport={{ once: true }}
+                transition={{ duration: 0.5 }}
+              >
+                <ProjectCard
+                  imgPath={project.imgPath}
+                  isBlog={false}
+                  title={project.title}
+                  description={project.description}
+                  ghLink={project.ghLink}
+                  demoLink={project.demoLink}
+                />
+              </motion.div>
+            </Col>
+          ))}
         </Row>
       </Container>
     </Container>
